+++ pytest_azurepipelines.py
@@ -7,7 +7,7 @@ import sys
 
 import pytest
 
-__version__ = "1.0.0.rc2"
+__version__ = "1.0.0"
 
 DEFAULT_PATH = "test-output.xml"
 DEFAULT_COVERAGE_PATH = "coverage.xml"
@@ -46,9 +46,15 @@ def pytest_addoption(parser):
 
 
 def pytest_configure(config):
-    nunit_xmlpath = config.getoption("--nunitxml")
-    if not nunit_xmlpath:
-        config.option.nunit_xmlpath = DEFAULT_PATH
+    if config.pluginmanager.has_plugin("pytest_nunit"):
+        nunit_xmlpath = config.getoption("--nunitxml")
+        if not nunit_xmlpath:
+            config.option.nunit_xmlpath = DEFAULT_PATH
+    else:
+        xmlpath = config.getoption("--junitxml")
+        if not xmlpath:
+            config.option.xmlpath = DEFAULT_PATH
+        # TODO set the family to xUnit 2
 
     # ensure coverage creates xml format
     if config.pluginmanager.has_plugin("pytest_cov"):
@@ -114,7 +120,12 @@ def try_to_inline_css_into_each_html_report_file(reportdir):
 
 
 def pytest_sessionfinish(session, exitstatus):
-    xmlpath = session.config.option.nunit_xmlpath
+    if session.config.pluginmanager.has_plugin("pytest_nunit"):
+        xmlpath = session.config.option.nunit_xmlpath
+        mode = "NUnit"
+    else:
+        xmlpath = session.config.option.xmlpath
+        mode = "xUnit"
 
     # This mirrors https://github.com/pytest-dev/pytest/blob/38adb23bd245329d26b36fd85a43aa9b3dd0406c/src/_pytest/junitxml.py#L368-L369
     xmlabspath = os.path.normpath(
@@ -135,8 +146,8 @@ def pytest_sessionfinish(session, exitstatus):
 
     if not session.config.getoption("no_docker_discovery"):
         print(
-            "##vso[results.publish type=NUnit;runTitle='{1}';publishRunAttachments=true;]{0}".format(
-                xmlabspath, description
+            "##vso[results.publish type={2};runTitle='{1}';publishRunAttachments=true;]{0}".format(
+                xmlabspath, description, mode
             )
         )
     else: